pyproject.toml
[tool.poetry]
name = "auto-mp3"
version = "0.1.0"
description = ""
authors = ["0xChloe <chloe@catastrophesoftware.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
qbittorrent-api = "^2025.2.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
